# Span marshalling using easyjson improves performance.