--- a/trace/format/trace_examples.go
+++ b/trace/format/trace_examples.go
@@ -1,0 +1,40 @@
+package traceformat
+
+// IndividualTrace is useful for
+const IndividualTrace = `{
+    "traceId": "string",
+    "name": "string",
+    "parentId": "string",
+    "id": "string",
+    "kind": "CLIENT",
+    "timestamp": 0,
+    "duration": 0,
+    "debug": true,
+    "shared": true,
+    "localEndpoint": {
+      "serviceName": "string",
+      "ipv4": "string",
+      "ipv6": "string",
+      "port": 0
+    },
+    "remoteEndpoint": {
+      "serviceName": "string",
+      "ipv4": "string",
+      "ipv6": "string",
+      "port": 0
+    },
+    "annotations": [
+      {
+        "timestamp": 0,
+        "value": "string"
+      }
+    ],
+    "tags": {
+      "additionalProp1": "string",
+      "additionalProp2": "string",
+      "additionalProp3": "string"
+    }
+  }`
+
+// ValidJSON is useful for testing
+const ValidJSON = "[" + IndividualTrace + "]"