--- a/trace/gobuild.toml
+++ b/trace/gobuild.toml
@@ -1,0 +1,2 @@
+[vars]
+  ignoreDirs = ["format"]