… sfxclient/httpsink_test.go
… trace/format/format_easyjson.go