benchmark                        old ns/op     new ns/op     delta
BenchmarkHTTPSink_AddSpans-8     413434        200635        -51.47%

benchmark                        old MB/s      new MB/s       speedup
BenchmarkHTTPSink_AddSpans-8     505952.18     3580817.47     7.08x

benchmark                        old allocs     new allocs     delta
BenchmarkHTTPSink_AddSpans-8     12             15             +25.00%

benchmark                        old bytes     new bytes     delta
BenchmarkHTTPSink_AddSpans-8     183952        76084         -58.64%